add admin action log middleware.
