+++ app/backend/app/Library/Log/AdminActionLogLibrary.php
@@ -48,81 +48,6 @@ class AdminActionLogLibrary
     ];
 
     /**
-     * Handle an incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  Closure  $next
-     * @return mixed
-     */
-    public function handle(Request $request, Closure $next)
-    {
-        if (self::isExcludePath($request->path())) {
-            return $next($request);
-        }
-
-        // POSTなど更新系のリクエストのみが対象
-        if (!in_array($request->method(), self::TARGET_REQUEST_METHOD_LIST, true)) {
-            return $next($request);
-        }
-
-        // $this->host = getmypid();
-        $requestDateTime = TimeLibrary::getCurrentDateTime();
-        $pid             = getmypid();
-
-        [
-            $uri,
-            $method,
-            $host,
-            $ip,
-            $contentType,
-            $plathome,
-            $headers,
-            $requestContent,
-        ] = self::getLogParameterByRequest($request);
-
-        // 処理速度の計測
-        $startTime = microtime(true);
-
-        $response = $next($request);
-
-        // 管理者の取得
-        $admin = Auth::guard('api-admins')->user();
-        // 管理者を取得出来ない場合
-        if (empty($admin) || empty($admin->id)) {
-            return $response;
-        }
-
-        $responseTime = (string)(microtime(true) - $startTime);
-        $memory = memory_get_usage();
-        $peakMemory = memory_get_peak_usage();
-
-        [$statusCode] = self::getLogParameterByResponse($response);
-
-        $description = '';
-
-        // log出力
-        self::outputLog(
-            $requestDateTime,
-            $uri,
-            $method,
-            $statusCode,
-            $responseTime,
-            $host,
-            $ip,
-            $contentType,
-            $headers,
-            $requestContent,
-            $plathome,
-            $description,
-            $pid,
-            $memory,
-            $peakMemory
-        );
-
-        return $response;
-    }
-
-    /**
      * check current path is log exclude path.
      *
      * @param string $path
